Implemented permission code on 'Controller/GroupAttributesController.php' in 'permission70' branch.

+++ htdocs/app/Controller/GroupAttributesController.php
@@ -18,23 +18,58 @@
 
 
 
+// Loads Util class.
+App::uses('Util', 'Utility');
+
+
+
 /**
- * Group Attributes
- *
  * @class GroupAttributesController
  *
  * @brief This class manages the group attributes.
  */
 class GroupAttributesController extends AppController
 {
+	/**
+	 * @var $components
+	 * This variable is used for include other conponents.
+	 */
+	var $components = array('Auth', 'Acl','Access');
+
+
+	/**
+	 * @var $helpers
+	 * This variable is used for include other helper file.
+	 */
+	var $helpers = array('Access');
+
+
+	/**
+	 * @method beforeFilter
+	 * This method executes method that we need to be executed before any other action.
+	 */
+	function beforeFilter()
+	{
+		parent::beforeFilter();
+	}
+
+
 
 	/**
 	 * @method index
-	 * @param  $groupId
 	 * This method is used to load list of group attributes with pagination.
+	 * @param  $groupId
 	 */
 	public function index($groupId)
 	{
+		// Get user group name.
+		$groupName = $this->Access->getGroupName($this->Session->read('User.ID'));
+		$this->set('groupName', $groupName);
+		// Check permission.
+		$permission = $this->Access->checkPermission('GroupAttributesController', 'View', $this->Session->read('User.ID'));
+		if (empty($permission)) {
+			throw new UnauthorizedException();
+		}
 		if (isset($groupId)) {
 			// Fetching data with pagination.
 			$this->paginate = array(
@@ -53,13 +88,19 @@ class GroupAttributesController extends AppController
 
 	/**
 	 * @method add
-	 * @param $groupId
 	 * This method is used to add group attributes.
+	 * @param $groupId
 	 */
 	public function add($groupId)
 	{
+		// Check permission.
+		$permission = $this->Access->checkPermission('GroupAttributesController', 'Add', $this->Session->read('User.ID'));
+		if (empty($permission)) {
+			throw new UnauthorizedException();
+		}
 		$this->set('groupId', $groupId);
-
+		$operators = Util::getAttributeOperators();
+		$this->set('operators', $operators);
 		if ($this->request->is('post')) {
 			$this->request->data['GroupAttribute']['Disabled'] = intval($this->request->data['GroupAttribute']['Disabled']);
 			$this->request->data['GroupAttribute']['GroupID'] = intval($this->request->params['pass'][0]);
@@ -69,9 +110,9 @@ class GroupAttributesController extends AppController
 			if ($this->GroupAttribute->validates()) {
 				// Saving data to table.
 				$this->GroupAttribute->save($this->request->data);
-				$this->Session->setFlash(__('Group attribute is saved succefully!', true), 'flash_success');
+				$this->Session->setFlash(__('Group attribute is saved successfully')."!", 'flash_success');
 			} else {
-				$this->Session->setFlash(__('Group attribute is not saved succefully!', true), 'flash_failure');
+				$this->Session->setFlash(__('Group attribute is not saved successfully')."!", 'flash_failure');
 			}
 		}
 	}
@@ -80,16 +121,22 @@ class GroupAttributesController extends AppController
 
 	/**
 	 * @method edit
+	 * This method is used to edit group attributes.
 	 * @param $id
 	 * @param $groupId
-	 * This method is used to edit group attributes.
 	 */
 	public function edit($id, $groupId)
 	{
+		// Check permission.
+		$permission = $this->Access->checkPermission('GroupAttributesController', 'Edit', $this->Session->read('User.ID'));
+		if (empty($permission)) {
+			throw new UnauthorizedException();
+		}
 		// Assigning group attribues values find by id to var.
 		$groupAttribute = $this->GroupAttribute->findById($id);
 		$this->set('groupAttribute', $groupAttribute);
-
+		$operators = Util::getAttributeOperators();
+		$this->set('operators', $operators);
 		if ($this->request->is('post')) {
 			$this->request->data['GroupAttribute']['Disabled'] = intval($this->request->data['GroupAttribute']['Disabled']);
 			$this->GroupAttribute->set($this->request->data);
@@ -97,9 +144,9 @@ class GroupAttributesController extends AppController
 				$this->GroupAttribute->id = $id;
 				//Saving data to the table.
 				$this->GroupAttribute->save($this->request->data);
-				$this->Session->setFlash(__('Attribute is saved succefully!', true), 'flash_success');
+				$this->Session->setFlash(__('Attribute was edited successfully')."!", 'flash_success');
 			} else {
-				$this->Session->setFlash(__('Attribute is not saved succefully!', true), 'flash_failure');
+				$this->Session->setFlash(__('Attribute was not saved successfully')."!", 'flash_failure');
 			}
 		}
 	}
@@ -108,24 +155,31 @@ class GroupAttributesController extends AppController
 
 	/**
 	 * @method remove
+	 * This method is used to delete group attributes.
 	 * @param $id
 	 * @param $groupId
-	 * This method is used to delete group attributes.
 	 */
 	public function remove($id, $groupId)
 	{
+		// Check permission.
+		$permission = $this->Access->checkPermission('GroupAttributesController', 'Delete', $this->Session->read('User.ID'));
+		if (empty($permission)) {
+			throw new UnauthorizedException();
+		}
 		if (isset($id)) {
 			// Deleting then redirecting to index function.
 			if ($this->GroupAttribute->delete($id)) {
 				$this->redirect('/group_attributes/index/'.$groupId);
-				$this->Session->setFlash(__('Attribute is removed succefully!', true), 'flash_success');
+				$this->Session->setFlash(__('Attribute was removed successfully')."!", 'flash_success');
 			} else {
-				$this->Session->setFlash(__('Attribute is not removed succefully!', true), 'flash_failure');
+				$this->Session->setFlash(__('Attribute was not removed successfully')."!", 'flash_failure');
 			}
 		} else {
 			$this->redirect('/group_attributes/index'.$userId);
 		}
 	}
 }
+
+
 
 // vim: ts=4
